+++ app/build.gradle.kts
@@ -38,7 +38,6 @@ android {
     kotlinOptions {
         jvmTarget = "1.8"
     }
-
 }
 
 dependencies {
