+++ codeAnalysis.sh
