--- data/src/main/java/com/android254/data/util/Converters.kt
@@ -1,15 +1,0 @@
-package com.android254.data.util
-
-import androidx.room.TypeConverter
-import kotlinx.datetime.Instant
-
-class InstantConverter {
-
-    @TypeConverter
-    fun longToInstant(value: Long?): Instant? =
-        value?.let(Instant::fromEpochMilliseconds)
-
-    @TypeConverter
-    fun instantToLong(instant: Instant?): Long? =
-        instant?.toEpochMilliseconds()
-}
