+++ presentation/src/main/java/com/android254/presentation/viewmodel/SessionViewModel.kt
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2022 DroidconKE
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.android254.presentation.viewmodel
 
 import androidx.lifecycle.ViewModel
@@ -5,7 +20,6 @@ import androidx.lifecycle.viewModelScope
 import com.android254.data.dao.SessionDao
 import com.android254.data.model.Session
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.launch
 import kotlinx.datetime.Clock
 import javax.inject.Inject
@@ -20,5 +34,4 @@ class SessionViewModel @Inject constructor(
     }
 
     val sessionsStream: List<Session> = sessionDao.fetchSessions()
-
 }
